+++ src/games/inss.rs
@@ -6,7 +6,7 @@ pub fn query(address: &str, port: Option<u16>) -> GDResult<game::Response> {
     let valve_response = valve::query(address, match port {
         None => 27131,
         Some(port) => port
-    }, SteamID::INSS.app(), None)?;
+    }, SteamID::INSS.as_app(), None, None)?;
 
     Ok(game::Response::new_from_valve_response(valve_response))
 }